fix: remove update_geo function that runs family query on the backend
I had originally left this in because I wanted to do the tidy up of the
geographies parameter separately, it seems that it causes an
intermittent bug that only gets triggered when you try and update one of
the basic properties. It is expecting the query to return one geo id but
it does not when we have families with multiple geos per the title so it
fails with :

Raises sqlalchemy.orm.exc.NoResultFound if the query selects no rows. Raises sqlalchemy.orm.exc.MultipleResultsFound if multiple object identities are returned, or if multiple rows are returned for a query that returns only scalar values as opposed to full identity-mapped entities.