feature/app-159-family-page-errors-in-the-admin-when-trying-to-update (#281)
* chore: update pyproject

* test: remove test - frontend has been updated to send this field always in the payload, so an error will raise if its not included

* test: update geo id validation mock call, we are using the get ids from values now

* fix: remove update_geo function that runs family query on the backend

I had originally left this in because I wanted to do the tidy up of the
geographies parameter separately, it seems that it causes an
intermittent bug that only gets triggered when you try and update one of
the basic properties. It is expecting the query to return one geo id but
it does not when we have families with multiple geos per the title so it
fails with :

Raises sqlalchemy.orm.exc.NoResultFound if the query selects no rows. Raises sqlalchemy.orm.exc.MultipleResultsFound if multiple object identities are returned, or if multiple rows are returned for a query that returns only scalar values as opposed to full identity-mapped entities.

* refactor: remove geo id from update service function

we are removing its use in the repository update function

* refactor: set geography and geo id to default to None

cleanup will be handled in a separate ticket

* tests: update tests and mocks accordingly

* pyproject version

* chore: add ports to docker dev yml file

allows users to access postico and other agents of the db in the dev container

* fix: pyproject version

* tests: add tests for multiple updates

* fix: failing integration tests

* tests: add a bit more clarity to tests

* fix: left geo in error when testing the test

---------

Co-authored-by: Osneil Drakes <osneildrakes@Osneils-MacBook-Pro.local>
Co-authored-by: Osneil Drakes <osneildrakes@Osneils-MBP.communityfibre.co.uk>